Extract and save abstract PDF after successful submissions
In test phases, users are required to submit an abstract in PDF format
that describes their approach. Add server-side processing to extract and
save the PDF to a subfolder of the submission folder after a successful
submission.

A reference to the PDF file ID is added to the submission's
'meta.abstractFileId' field.